+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Do not halt bulk-IN on a passed short read (BOT 6.7.2 case 5) (https://github.com/apohrebniak/usbd-storage/pull/23)
 - Rust 2024 edition (https://github.com/apohrebniak/usbd-storage/pull/22)
 - Fix unsoundness issue in internal use of `BulkOnly`s buffer (https://github.com/apohrebniak/usbd-storage/pull/21)